feat: moved some utils to api-utils.ts and typed them  feat: changed api.js to api.ts
